refractor: pass aframe values as template literal strings
Renames model.js component to volume.js
